requirements.txt
langchain>=1.2.12
langchain-core>=1.2.19
langchain-community>=0.4.1
langchain-openai>=0.3.0
langchain-voyageai>=0.1.3
langchain-text-splitters>=1.1.1
langgraph>=1.1.2
faiss-cpu>=1.13.2
pydantic>=2.12.4
python-dotenv>=1.1.0
beautifulsoup4>=4.13.5
requests>=2.32.5
streamlit>=1.51.0
wikipedia>=1.4.0
pypdf>=6.8.0